Refactor FoldingManager and FoldingSection classes to improve event handling and encapsulate collapsed line section validation
